Move charts components to UI folder
