+++ src/pages/search/outlets/landscape/connectivity/CurrentPAConnectivityController.ts
@@ -1,4 +1,4 @@
-import { SmallBarsData } from "pages/search/shared_components/charts/SmallBars";
+import { SmallBarsData } from "pages/search/ui/charts/SmallBars";
 import { DPC } from "pages/search/types/connectivity";
 import formatNumber from "pages/search/utils/format";
 import { SmallBarTooltip } from "pages/search/types/charts";
