+++ src/pages/search/outlets/landscape/humanFootprint/CurrentFootprint.tsx
@@ -17,7 +17,7 @@ import {
   currentHFCategories,
 } from "pages/search/types/humanFootprint";
 import { textsObject } from "pages/search/types/texts";
-import LargeStackedBar from "pages/search/shared_components/charts/LargeStackedBar";
+import LargeStackedBar from "pages/search/ui/charts/LargeStackedBar";
 import { wrapperMessage } from "pages/search/types/charts";
 import { CurrentFootprintController } from "pages/search/outlets/landscape/humanFootprint/CurrentFootprintController";
 import { ShapeLayer } from "pages/search/types/layers";
