+++ src/pages/search/ui/charts/BarItem.tsx
