+++ src/pages/search/ui/charts/Lines.tsx
@@ -3,7 +3,7 @@ import { Point, ResponsiveLine } from "@nivo/line";
 import { CartesianMarkerProps } from "@nivo/core";
 
 import formatNumber from "pages/search/utils/format";
-import withMessageWrapper from "pages/search/shared_components/charts/withMessageWrapper";
+import withMessageWrapper from "pages/search/ui/charts/withMessageWrapper";
 
 interface LinesData {
   label: string;
