+++ src/pages/search/ui/charts/Pie.tsx
@@ -3,7 +3,7 @@ import React from "react";
 
 import formatNumber from "pages/search/utils/format";
 import { lightenColor, darkenColor } from "pages/search/utils/colorUtils";
-import withMessageWrapper from "pages/search/shared_components/charts/withMessageWrapper";
+import withMessageWrapper from "pages/search/ui/charts/withMessageWrapper";
 
 interface Props {
   data: Array<PieData>;
